% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{assign_stars}
\alias{assign_stars}
\title{Assign stars based on p-values.}
\usage{
assign_stars(pval)
}
\arguments{
\item{pval}{a number}
}
\description{
This gives a string with stars corresponding to conventional alpha values (0.001, 0.01, 0.05) based on the value of the input.
}
