% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{geom_ccol}
\alias{geom_ccol}
\title{Cody's preferred geom_col}
\usage{
geom_ccol(..., bar_width = 0.6, position = position_dodge(),
  color = "black", size = 0.75)
}
\arguments{
\item{bar_width}{Width of the column. The same width should be provided to geom_cerrorbar, if used, so that dodging is aligned.}

\item{position}{see ggplot2::geom_col}

\item{color}{see ggplot2::geom_col}

\item{size}{see ggplot2::geom_col}
}
\description{
This function is a wrapper for ggplot2::geom_col. It specifies some stylistic elements so they don't need to be typed out each time. It is meant to be used with geom_cerrorbar.
}
