% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{geom_cerrorbar}
\alias{geom_cerrorbar}
\title{Cody's preferred geom_errorbar}
\usage{
geom_cerrorbar(..., bar_width = 0.6, position = position_dodge(width =
  bar_width), color = "black", width = bar_width * 0.5, size = 0.75)
}
\arguments{
\item{bar_width}{Width of the geom that cerrorbar will be added to.}

\item{position}{see ggplot2::geom_errorbar}

\item{color}{see ggplot2::geom_errorbar}

\item{width}{By default, 0.5*bar_width.}

\item{size}{see ggplot2::geom_errorbar}
}
\description{
This calls ggplot2:geom_errorbar with some shortcuts for use with the other codyverse functions.
}
