% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mean_sem}
\alias{mean_sem}
\title{Mean and SEM Calculator}
\usage{
mean_sem(x)
}
\arguments{
\item{x}{}
}
\description{
This gives a dataframe containing y, ymin, and ymax values corresponding to the mean, mean-SEM, and mean+SEM of the input data. It can be called as the fun.data from ggplot2::stat_summary
}
