% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{scale_y_prettylog10}
\alias{scale_y_prettylog10}
\title{Pretty log 10}
\usage{
scale_y_prettylog10(...)
}
\arguments{
\item{...}{Any arguments to pass to ggplot2::scale_y_log10}
}
\description{
Calls ggplot2::scale_y_log10 with labels for "pretty" exponents. Recommend to also add annotation_logticks(sides = "l")
}
