% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sem}
\alias{sem}
\title{Standard Error of the Mean}
\usage{
sem(x)
}
\arguments{
\item{x}{Array of which you want to calculate the SEM.}
}
\value{
Returns a numeric SEM.
}
\description{
This function returns the standard error of the mean (SEM) of a given numeric vector.The SEM defined as the stadard deviation (SD) divided by the square root of the count of observations. In other words, SD(x)/sqrt(n).
}
\examples{
sem(c(3, 3, 3.5, 4))
sem(iris$Sepal.Length)
}
