% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{theme_cody}
\alias{theme_cody}
\title{Cody's preferred theme for ggplot2}
\usage{
theme_cody(base.size = 14)
}
\description{
This builds upon theme_bw from ggplot2 to make figures look the way I prefer. Inspect the source for more info on this.
}
\examples{
ggplot(iris, aes(x=Sepal.Length, y = Sepal.Width)) +
geom_point() +
theme_cody()
}
